Update ko.json
Translated Rest of File settings and Appearance settings.
Took some liberty and made some wording decisions. For the "Vault" I wasn't sure which word to use so I used "보관함" which actually means more like a "Storage." But a direct translation of Vault is a "금고" which I'm not sure if it would be understood as a "Knowledge Storage." Feel like the Nuance is a bit different, where in English we can kind of see "Vault" as a place to keep stuff safe. But I think in Korean "금고" is seen more as just storage for Money only.... This might be my opinion so I can't be sure. If you guys want to give the idea that Obsidian folder is a storage for keeping knowledge safe like a "vault," then "금고" might still work, as "보관함" is also pretty vague.... I just couldn't think of a good word for Vault.

For Appearance, I just named it "테마" which actually means "Themes," The Direct translation of appearance would be "외관" but that's a word used for appearance of physical objects, and I don't think I've ever heard people use that term for "Software appearances." It also sounds really weird.... The word "테마" or "Theme" gives a more easier understanding that it's an option for changing the appearance.